restructuring participant manager files
